Post must have an author to be publicated
