+++ db/migrate/20220330024157_add_user_id_to_posts.rb
@@ -1,0 +1,5 @@
+class AddUserIdToPosts < ActiveRecord::Migration[6.0]
+  def change
+    add_reference :posts, :user, index: true
+  end
+end
